RxJulia Guide

This package is an implementation of the ReactiveX reactive programming model in Julia.

Just like other libraries in the ReactiveX ecosystem, the basic Observable and Subject types exist (the latter is called Reactor here), together with many operators for constructing new observables from existing ones.

RxJulia takes advantage of Julia's macro and type systems to create a domain-specific language (DSL) useful in reactive programming: see Pipelines for more details.

Contents

RxJulia Guide

Contents

Pipelines

Kernel

Functions

Macros

Types

Constants

Operators

Functions

Index

Pipelines

One area that is unique to RxJulia is the use of macros to create pipelines or chains of Reactor instances in code. Specifically the @rx macro makes this possible.

events = @rx do
  observableFactory1
  reactorFactory1
  reactorFactory2
  # more....
end

for event in events
  # do something with the event.
end

Note that intermediate stages in a pipeline are Reactors, capable of both being an Observer and an Observable. Both the beginning and the end of the pipeline should be an Observable (of course, another Reactor is fine too). A Reactor is the analogue of a Subject in the ReactiveX model.

When using @rx, a special observable, a Collector, is added to the end of the chain, resulting in an object suitable for iteration. Iteration ends when a CompletedEvent or an ErrorEvent are encountered. Thus, it is possible to iterate over the events generated from the created pipeline.

evt = @rx(() -> 1)

Collector(Channel{Any}(sz_max:32,sz_curr:2))

Each statment inside the @rx block is actually an expression that should be a factory for creating an Observable or Reactor, and each line subscribes to the events generated by the line above it, with subscribe!.

Fortunately, nearly any value will work as an Observable: scalar values such as strings, numbers, booleans, symbols, etc., all result in an Observable that produces only a single value before completion.

evts = @rx() do
  1
end
for evt in evts
  println("Event: $evt")
end

Event: 1

Values that are iterable will result in a stream of values passed to the downstream observers, completing when iteration ends.

evts = @rx() do
  [1 2 3]
end
for evt in evts
  println("Event: $evt")
end

Event: 1
Event: 2
Event: 3

Chaining occurs naturally within the pipeline (see select for its meaning in this context):

evts = @rx() do
  [1, 2, 3, 4, 5, 6]
  select(isodd)
end
println([evt for evt in evts])

[1, 3, 5]

Each pipeline created with @rx is itself an Observable.

evts1 = @rx() do
  [1 2 3]
end

evts2 = @rx() do
  evts1
end

for evt in evts2
  println("Event: $evt")
end

Event: 1
Event: 2
Event: 3

Kernel

The kernel is the fundamental implementation of the reactive pattern in Julia, independent of any operators that leverage the package's capabilities.

Functions

RxJulia.dispatch — Method

dispatch(fn::Function)::Reactor = Reactor(fn)

Return a Reactor around the provided function which takes Observers and an Event, taking action as appropriate, and producing an Observable that can also be an Observer of other Observables.

RxJulia.events — Function

events(sz = 32)

Return a Collector to accumulate observed Events

RxJulia.notify! — Method

notify!(observers, event)

Notify Observer by invoking the block on each

RxJulia.onEvent — Method

onEvent(observer, event)

Deliver an Event to the Observer

RxJulia.react — Method

react(fn::Function)::Reactor

Return a [Reactor](@ref) around the provided function which takes Observers and a value, taking action as appropriate, and producing an Observable that can also be an Observer of other Observables. [CompletedEvent]](@ref)s and ErrorEvents are passed on to Observers, while [ValueEvent]](@ref)s are passed to the supplied function. The supplied function takes [Observers] and a value as arguments.

RxJulia.subscribe! — Method

Treat any value as an Observable, and subscribe the Observer to it; the value will be emitted once with a ValueEvent, then a CompletedEvent will be emitted. If the value is iterable (e.g., applicable(iterate, value) returns true), then emit over each value returned by the iterable and pass on to the Observer, concluding with a CompletedEvent. Note that strings are treated as an atomic value and not an iterable.

# works for booleans
julia> evts = @rx(() -> true)
Collector(Channel{Any}(sz_max:32,sz_curr:2))

julia> collect(evts)
1-element Array{Any,1}:
 true

# works for booleans
julia> evts = @rx( ()-> 1 )
Collector(Channel{Any}(sz_max:32,sz_curr:2))

julia> collect(evts)
1-element Array{Any,1}:
 1

# works for strings
julia> evts = @rx( ()-> "foo" )
Collector(Channel{Any}(sz_max:32,sz_curr:2))

julia> collect(evts)
1-element Array{Any,1}:
 "foo"

# works for iterables
julia> evts = @rx( ()-> [1 2 3] )
Collector(Channel{Any}(sz_max:32,sz_curr:4))

julia> collect(evts)
3-element Array{Any,1}:
 1
 2
 3

RxJulia.subscribe! — Method

subscribe!(observers::Observers, observer)

Add an Observer to an existing set of Observers, return the Observer

RxJulia.subscribe! — Method

subscribe!(observable::Observable, observer)

Subscribe an Observer to the given Observable, return the Observer

RxJulia.chain! — Method

chain!(observable, observer)

Subscribe the Observer to the Observable using subscribe!, and return the Observable.

RxJulia.complete! — Method

complete!(observers)

Emit a CompletedEvent](@ref) to Observers

RxJulia.error! — Method

error!(obserers, err)

Emit an ErrorEvent to Observers

RxJulia.subscribe_iterable! — Method

subscribe_iterable!(iterable, observer)

Treat an iteraable as an Observable, emitting each of its elements in turn in a ValueEvent, and concluding with a CompletedEvent.

RxJulia.subscribe_value! — Method

subscribe_value!(value, observer)

Treat a value as an Observable, emitting the value followed by a CompletedEvent

Macros

RxJulia.@rx — Macro

rx(blk)

Given a do-block where each statement is an Observable, subscribe! each in sequence to the one proceeding. Return an object that one can use to iterate over the events from the last Observable in the block.

Example:

results = @rx() do
count
filter(:even)
end
printlin([result for result in results])

Types

RxJulia.Collector — Type

An Observer that accumulates events on a Channel, which then may be retrieved for iteration

RxJulia.CompletedEvent — Type

An Event signifying no more events are available from the originating Observable.

RxJulia.ErrorEvent — Type

An Event signifying the originating Observable encountered an error; no more events will be delivered after events of tyie.

RxJulia.Observable — Type

An Observable is a source of Events, with Observers subscribe!ing to the Observable in order to receive those events.

See Observable in ReactiveX documentation

RxJulia.Observer — Type

An Observer is a receiver of Events via onEvent

RxJulia.Reactor — Type

Reactor() = Reactor(notify!)
Reactor(fn) = Reactor(fn, [])

A Reactor is an Observer that is also useful for building Observers that are Observable as well.

A Reactor is the analogue of a Subject in the ReactiveX model.

RxJulia.ValueEvent — Type

An Event containing a value to deliver to an Observable.

RxJulia.Observers — Type

Observers is a collection of subscribed Observers

Constants

RxJulia.Event — Constant

An Event is any of the following:

A value, encapsulated as ValueEvent

A completion event, encapsulated as CompletedEvent

An error, encapsulated as ErrorEvent

Operators

Operators are useful functions for producing Reactors that receive events and decide whether to transform them or skip emitting values to downstream observers.

Functions

RxJulia.cut — Method

cut(n)

Drop the last n events observed, emitting all others that precede them. If less than n events observed, emit nothing.

RxJulia.distinct — Method

distinct()

Only emit unique values observed

RxJulia.drop — Method

drop(n)

Drop the first n events observed, emitting all others that precede them. If less than n events observed, emit nothing.

RxJulia.fmap — Method

fmap(fn::Function)

Apply function to all observed values, and emit the result. Functional equivalent of map.

RxJulia.keep — Method

keep(n)

Keep only the last n values, discarding the rest. If less than n values observed, emit only values observed.

RxJulia.merge — Method

merge(observables...)

Produce an Observable that emits from the supplied Observables, until either all have emitted a CompletedEvent or one of them emits an ErrorEvent

RxJulia.reject — Method

reject(fn::Function)::Reactor

Apply a filter function such that only values for which the function returns false will be passed onto Observers

RxJulia.select — Method

select(fn::Function)

Apply a filter function such that only values for which the function returns true will be passed onto Observers

RxJulia.span — Method

span(m, n)

Emit only the integers in the range of m to n, inclusive

RxJulia.take — Method

take(n)

Take only the first n values, discarding the rest. If less than n values observed, emit only values observed.

RxJulia.zip — Method

zip(observables...)

Given one or more Observables, read one value at a time from each and collect into a Tuple, stopping when either there is an ErrorEvent or any CompletedEvent

Index

RxJulia.Event

RxJulia.Collector

RxJulia.CompletedEvent

RxJulia.ErrorEvent

RxJulia.Observable

RxJulia.Observer

RxJulia.Observers

RxJulia.Reactor

RxJulia.ValueEvent

RxJulia.chain!

RxJulia.complete!

RxJulia.cut

RxJulia.dispatch

RxJulia.distinct

RxJulia.drop

RxJulia.error!

RxJulia.events

RxJulia.fmap

RxJulia.keep

RxJulia.merge

RxJulia.notify!

RxJulia.onEvent

RxJulia.react

RxJulia.reject

RxJulia.select

RxJulia.span

RxJulia.subscribe!

RxJulia.subscribe!

RxJulia.subscribe!

RxJulia.subscribe_iterable!

RxJulia.subscribe_value!

RxJulia.take

RxJulia.zip

RxJulia.@rx